Fix failing tests by seperating lotterystudentregauth from ajaxstudentreg
